Add files via upload
Developed the user login mechanism that authenticates users by scanning the DB and comparing the entered and saved passwords. The administrator portal is almost complete. It allows the creation and deletion of users. It allows the creation and deletion of courses. The web portion of the login page is finished and the administrator portal is mostly done. Web portion is in JSP files, web input processing is done by serverlets in Java files, and DB logic is in Java files.